Removing default_parameter.py and transfer contents utils/training/read_config.py
